api/requirements.txt
# Arquivo api/requirements.txt
Flask
psycopg2
kafka-python